auth: move CAT4MOQ issuer into standalone moqx-issuer binary
The relay now only verifies tokens; signing lives in a separate binary and
library, not moqx_core.

- moqx-issuer: standalone CLI (src/auth/IssuerMain.cpp), same flags/output as
  the former issue-cat-token subcommand
- moqx_issuer_lib: AuthTokenIssuer + signGrants (tokenFromGrants/toCatapultMatch
  moved out of Auth.cpp)
- moqx_hmac_key: shared deriveHmacKey so verify and sign can't drift on HKDF
  salt/info
- moqx relay: drop sign()/signForTest() and the issue-cat-token subcommand
- docker: entrypoint routes issue-cat-token to moqx-issuer
